backend/requirements.txt
asgiref==3.3.4
Django==3.2.2
django-cors-headers==3.13.0
djangorestframework==3.14.0